requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
alembic==1.13.1
aiosqlite==0.20.0
pydantic==2.8.2
email-validator==2.1.1
loguru==0.7.2
pytest==8.3.2
pytest-asyncio==0.23.7
pytest-xdist==3.6.1
pytest-cov==5.0.0
httpx==0.27.0
requests==2.32.3
allure-pytest==2.13.5
faker==26.0.0
starlette==0.37.2
